# FW: move the between-test checks on MCE and thermal events to mce_check