This removes the parsing of /proc/interrupts on Linux between each test and when the "mce_check" test is run. In any case, the product of those checks was only comments, not properly structured data (because we introduced them for the TAP output format, before we had YAML), which means many tools never caught or stored them at all. Though as a consequence the test running in a child process, its update of the counts won't be committed, so when the test is run again, it'll compare against the original numbers from the tool's start.

The problem with /proc/interrupts is that it increases in size linearly with the number of cores.

This improves start-up time a bit if `--disable=mce_check` is passed. For `perf stat -r500 $objdir/opendcdiag --disable=mce_check --selftests -o - -e selftest_skip`:

| Machine     | Before (ms) | After (ms) |
|-------------|-------------|------------|
| 14c SKX     |       3.94  |      3.46  |
| 2x28c CLX   |      36.40  |     27.83  |
| 2x48c SPR   |     123.86  |     80.00  |

And 1000 runs of `selftest_pass` (not including the start up time above):

| Machine     | Before (ms) | After (ms) |
|-------------|-------------|------------|
| 14c SKX     |       6.83  |      6.52  |
| 2x28c CLX   |      21.45  |     13.91  |
| 2x48c SPR   |      63.73  |     17.23  |

[ChangeLog][mce_check] This test now has a "sticky" behavior: if a problem is found, it is reported every time the test is run, instead of only the first time the problem was detected.